feat: navigator to hud scene
