+++ client/src/three/scenes/HUDScene.ts
@@ -1,18 +1,35 @@
 import * as THREE from "three";
 import { MapControls } from "three/examples/jsm/controls/MapControls";
 import { SceneManager } from "../SceneManager";
+import { Navigator } from "../components/Navigator";
+import { GUIManager } from "../helpers/GUIManager";
 
 export default class HUDScene {
   private scene: THREE.Scene;
   private camera: THREE.OrthographicCamera;
   private sceneManager: SceneManager;
   private controls: MapControls;
+  private GUIFolder: any;
+  private navigator: Navigator;
 
   constructor(sceneManager: SceneManager, controls: MapControls) {
     this.scene = new THREE.Scene();
     this.sceneManager = sceneManager;
+    this.controls = controls;
     this.camera = this.createOrthographicCamera();
-    this.initHUD();
+    this.GUIFolder = GUIManager.addFolder("HUD");
+
+    this.navigator = new Navigator(this.scene, this.controls, this.GUIFolder);
+    const navigatorParams = { col: 269, row: 153 };
+    this.GUIFolder.add(navigatorParams, "col").name("Col");
+    this.GUIFolder.add(navigatorParams, "row").name("Row");
+    this.GUIFolder.add(
+      {
+        setNavigationTarget: () => this.navigator.setNavigationTarget(navigatorParams.col, navigatorParams.row),
+      },
+      "setNavigationTarget",
+    ).name("Navigate to Col Row");
+    this.GUIFolder.close();
   }
 
   private createOrthographicCamera(): THREE.OrthographicCamera {
@@ -30,16 +47,6 @@ export default class HUDScene {
     return camera;
   }
 
-  private initHUD() {
-    // Add HUD elements here
-    const arrowGeometry = new THREE.ConeGeometry(0.2, 0.5, 32);
-    const arrowMaterial = new THREE.MeshBasicMaterial({ color: 0xffff00 });
-    const arrow = new THREE.Mesh(arrowGeometry, arrowMaterial);
-    arrow.position.set(-4, -2, 0);
-    arrow.rotation.z = Math.PI / 4;
-    this.scene.add(arrow);
-  }
-
   getScene(): THREE.Scene {
     return this.scene;
   }
@@ -50,6 +57,7 @@ export default class HUDScene {
 
   update(deltaTime: number) {
     // Update HUD elements if needed
+    this.navigator.update();
   }
 
   onWindowResize(width: number, height: number) {
@@ -60,5 +68,9 @@ export default class HUDScene {
     this.camera.top = frustumSize / 2;
     this.camera.bottom = frustumSize / -2;
     this.camera.updateProjectionMatrix();
+  }
+
+  setNavigationTarget(col: number, row: number) {
+    this.navigator.setNavigationTarget(col, row);
   }
 }
